fix: default example parser with alias
closes #806
